rag-cli/pyproject.toml
[project]
name = "rag-cli"
version = "0.1.0"
description = "A high-precision, framework-free RAG CLI tool powered by Gemini + ChromaDB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "pypdf>=4.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
